Example of how to filter the CSV
This script shows you how you can eliminate and filter the Studentinfo CSV for other purposes you may have or need. This is helpful if you only want to create accounts for your a particular group of students, but do not want to make changes to the base studentinfo csv.